packet train

+++ net/ipv4/tcp_cubic.c
@@ -428,8 +428,8 @@ static void hystart_update(struct sock *sk, u32 delay)
 				u32 packet_pair_time = now - last_ack;
 				if (packet_pair_time > 0)
 				{
-					u64 est_packet_pair_bd = (1500 * 8 * 1000000 / packet_pair_time);
-					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Packet pair time %u, Est. bandwidth %lld Mb/s\n",
+					unsigned long long est_packet_pair_bd = (1500 * 8 * 1000000 / packet_pair_time);
+					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Packet pair time %u, Est. bandwidth %llu Mb/s\n",
 						port, round_id, now, ca->round_start, packet_pair_time, est_packet_pair_bd);
 				}
